refactor: Rename channels to channel
