+++ crates/scheduler/src/channel/interval.rs
