+++ crates/scheduler/src/lib.rs
@@ -35,7 +35,7 @@ pub enum Error {
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
 
-pub mod channels;
+pub mod channel;
 pub mod schedule;
 pub mod scheduler;
 pub mod tasks;
